moving to tsx
