+++ package.json
@@ -32,15 +32,20 @@
   ],
   "devDependencies": {
     "typescript": "2.6.1",
+    "react": "16.3.1",
     "semantic-ui-less": "2.2.14",
     "@types/bluebird": "2.0.33",
     "@types/marked": "0.3.0",
     "@types/node": "8.0.53",
-    "webfonts-generator": "^0.4.0"
+    "webfonts-generator": "^0.4.0",
+    "@types/jquery": "3.2.16",
+    "@types/react": "16.0.25",
+    "@types/react-dom": "16.0.3",
+    "@types/web-bluetooth": "0.0.4"
   },
   "dependencies": {
-    "pxt-common-packages": "6.4.13",
-    "pxt-core": "5.5.18"
+    "pxt-common-packages": "6.16.5",
+    "pxt-core": "5.24.3"
   },
   "scripts": {
     "test": "node node_modules/pxt-core/built/pxt.js travis"
